Fix deleting customer post-booking causing error

+++ src/services/CustomerService.java
@@ -1,8 +1,10 @@
 package services;
 
 import dao.impl.CustomerDAOImpl;
+import db.TransactionManager;
 import models.Customer;
 
+import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
 
@@ -32,11 +34,32 @@ public class CustomerService {
     }
 
     public Optional<Customer> deleteCustomer(int customerId) {
+        Optional<Customer> opt;
+
+        // Get customer
         Customer customer = customerDAO.getCustomerById(customerId);
 
-        if (customerDAO.deleteCustomer(customer) > 0) {
-            return Optional.of(customer);
-        } else {
+        try {
+            TransactionManager.begin();
+
+            // Mark rooms booked by customer as available
+            if (!customerDAO.removeCustomerBookings(customer)) {
+                TransactionManager.rollback();
+                return Optional.empty();
+            }
+
+            // Delete customer
+            if (customerDAO.deleteCustomer(customer) > 0) {
+                TransactionManager.commit();
+                return Optional.of(customer);
+            } else {
+                TransactionManager.rollback();
+                return Optional.empty();
+            }
+        } catch (SQLException e) {
+            TransactionManager.rollback();
+            return Optional.empty();
+        } catch (Exception ignored) {
             return Optional.empty();
         }
     }
